% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etapa1_consorégionale.R
\name{Etape1_conso_régionale}
\alias{Etape1_conso_régionale}
\title{Calcule la consommation moyenne de produits par tête et par région}
\usage{
Etape1_conso_régionale(base_brute, df_conv, membres, adresse)
}
\arguments{
\item{base_brute}{data.frame. Base brute de consommation (ex: laitier)}

\item{df_conv}{data.frame. Table de conversion avec poids}

\item{membres}{data.frame. Table des membres du ménage pour calculer la taille}

\item{adresse}{data.frame. Table avec les variables de région et milieu}
}
\value{
Un tableau avec région, total consommé, taille totale, consommation moyenne par tête
}
\description{
Cette fonction traite une base brute de consommation (comme `laitier`),
applique les conversions, fusionne avec la taille des ménages et les infos régionales,
et retourne la consommation moyenne par personne et par région.
}
\examples{
conso_par_tete_region(laitier, df_conv, membres, adresse)
}
